Move helper methods to new class and do cleanup
The thread factories can be combined to one factory
